# Lib zebra h cleanup