--- a/bfdd/bfd_packet.c
+++ b/bfdd/bfd_packet.c
@@ -12,6 +12,7 @@
  */
 
 #include <zebra.h>
+#include <sys/ioctl.h>
 
 #ifdef GNU_LINUX
 #include <linux/filter.h>