--- a/lib/vrf.c
+++ b/lib/vrf.c
@@ -5,6 +5,7 @@
  */
 
 #include <zebra.h>
+#include <sys/ioctl.h>
 
 #include "if.h"
 #include "vrf.h"