--- a/lib/zebra.h
+++ b/lib/zebra.h
@@ -83,11 +83,6 @@
 #endif
 #endif
 
-#ifdef CRYPTO_OPENSSL
-#include <openssl/evp.h>
-#include <openssl/hmac.h>
-#endif
-
 #include "openbsd-tree.h"
 
 #include <netinet/in.h>