--- a/lib/zebra.h
+++ b/lib/zebra.h
@@ -81,7 +81,6 @@
 #include "openbsd-tree.h"
 
 #include <netinet/in.h>
-#include <netinet/in_systm.h>
 #include <netinet/ip.h>
 #include <netinet/tcp.h>