--- a/lib/zebra.h
+++ b/lib/zebra.h
@@ -22,7 +22,6 @@
 #include <signal.h>
 #include <string.h>
 #include <pwd.h>
-#include <grp.h>
 #ifdef HAVE_STROPTS_H
 #include <stropts.h>
 #endif /* HAVE_STROPTS_H */