--- a/lib/zlog_live.c
+++ b/lib/zlog_live.c
@@ -5,6 +5,8 @@
 
 #include "zebra.h"
 
+#include "frrsendmmsg.h"
+
 #include "zlog_live.h"
 
 #include "memory.h"