--- a/nhrpd/linux.c
+++ b/nhrpd/linux.c
@@ -7,6 +7,7 @@
 
 #include <errno.h>
 #include <linux/if_packet.h>
+#include <sys/ioctl.h>
 
 #include "nhrp_protocol.h"
 #include "os.h"