--- a/ospfd/ospf_auth.c
+++ b/ospfd/ospf_auth.c
@@ -6,6 +6,11 @@
 
 #include <zebra.h>
 
+#ifdef CRYPTO_OPENSSL
+#include <openssl/evp.h>
+#include <openssl/hmac.h>
+#endif
+
 #include "linklist.h"
 #include "if.h"
 #include "checksum.h"