--- a/sharpd/sharp_logpump.c
+++ b/sharpd/sharp_logpump.c
@@ -5,6 +5,7 @@
  */
 
 #include <zebra.h>
+#include <sys/resource.h>
 
 #include "vty.h"
 #include "command.h"