Zebra.h has become a bit of a dumping ground for system headers.  Let's make a single pass through zebra.h and properly locate stuff where it belongs.  Especially for includes that are only used by 1-2 files.

Initial results of compile time on my machine actually surprised me.

I run a make -j32

Before timings:
```
   Executed in   13.12 secs    fish           external
   usr time  208.02 secs  360.00 micros  208.02 secs
   sys time   68.33 secs  179.00 micros   68.33 secs
```

After Timings:

```
Executed in   10.92 secs    fish           external
   usr time  158.24 secs  328.00 micros  158.24 secs
   sys time   50.81 secs  162.00 micros   50.81 secs
```